refactor CoverageServiceImpl
